refactor(triage): split tests into triage/tests.rs to satisfy the 1500-line ratchet
CI's file-size gate rejected triage.rs at 1559 lines. Move the unit tests
into a child module (src/triage/tests.rs), mirroring the existing
pipeline.rs / pipeline/tests.rs layout, which puts triage.rs at 956 lines.

Retighten the baseline for the two grandfathered files this branch grew
(pipeline.rs +8 for the headless chat veto, pipeline/tests.rs +34 for the
new headless routing test). The zai.rs 1520 -> 1524 bump records growth
that landed on main in #1202 without a baseline update — main's ratchet
gate is red on that today; this makes the recorded ceiling match reality.
