+++ scripts/file-size-baseline.txt
@@ -25,10 +25,10 @@
 1849 stella-fleet/src/fleet.rs
 1628 stella-model/src/anthropic/tests.rs
 2060 stella-model/src/openai.rs
-1520 stella-model/src/zai.rs
+1524 stella-model/src/zai.rs
 1773 stella-model/src/zai/tests.rs
-3381 stella-pipeline/src/pipeline.rs
-2440 stella-pipeline/src/pipeline/tests.rs
+3389 stella-pipeline/src/pipeline.rs
+2474 stella-pipeline/src/pipeline/tests.rs
 2784 stella-protocol/src/event.rs
 2078 stella-store/src/lib.rs
 2261 stella-store/src/tests.rs
